refactor: create abstract action class
